Move types definition to models folder
