+++ frontend/src/models/editor/layers/layerConfigurationProps.ts
@@ -1,0 +1,8 @@
+import { LayerUpdateCallback } from "./layer";
+
+export type LayerConfigurationProps = {
+  name: string;
+  isSelected: boolean;
+  isVisible: boolean;
+  updateLayer: (callback: LayerUpdateCallback) => void;
+};
