+++ frontend/src/models/editor/layers/layerManagementProps.ts
@@ -1,0 +1,6 @@
+import { Layer, LayerId, LayerUpdateCallback } from "./layer";
+
+export type LayersManagementProps = {
+  layers: Layer[];
+  updateLayer: (id: LayerId, callback: LayerUpdateCallback) => void;
+};
