+++ frontend/src/models/editor/layers/layerProps.ts
@@ -1,0 +1,16 @@
+import { Vector2d } from "konva/lib/types";
+import { LayerId } from "./layer";
+import { Line } from "./line";
+
+export interface LayerProps {
+  id: LayerId;
+  position: Vector2d;
+  rotation: number;
+  scale: Vector2d;
+  image: HTMLImageElement;
+  isVisible: boolean;
+  lines: Line[];
+
+  onDragEnd: any;
+  onTransformEnd: any;
+}
