+++ frontend/src/models/editor/utils/CircularBuffer.ts
